solution/pyproject.toml
[project]
name = "solution"
version = "0.1.0"
description = "Final task for PROD-2025 (backend)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.36.24",
    "celery>=5.4.0",
    "django-storages[s3]>=1.14.5",
    "django>=5.1.6",
    "django-ninja>=1.3.0",
    "flake8>=7.1.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentry-sdk>=2.22.0",
    "yandex-cloud-ml-sdk>=0.3.1",
    "django-s3-storage>=0.15.0",
]
